requirements.txt
# Core LangChain + LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0
langchain-neo4j>=0.1.1
langgraph>=0.2.0

# PostgreSQL + pgvector (replaces chromadb + rank-bm25)
psycopg2-binary>=2.9.9
pgvector>=0.3.0
langchain-postgres>=0.0.9

# RAG + document processing
pypdf>=4.0.0
docx2txt>=0.8
python-docx>=1.1.0
rapidfuzz>=3.6.0

# Docling — advanced document extraction engine (IBM, MIT license)
# Handles PDF, DOCX, PPTX, XLSX, images; layout-aware; table extraction; clean Markdown output
# Install separately if not needed: pip install docling
docling>=2.0.0
graphrag>=3.0.8
pyarrow>=16.0.0

# OCR support (optional — CPU; swap paddlepaddle → paddlepaddle-gpu for CUDA)
paddlepaddle>=2.6.0
paddleocr>=2.7.0
pymupdf>=1.24.0
pillow>=10.0.0

# Observability
langfuse>=2.0.0

# CLI / utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.12.0
rich>=13.0.0

# API gateway
fastapi>=0.115.0
uvicorn>=0.30.0
httpx[http2]>=0.27.0
python-multipart>=0.0.9
mcp>=1.0.0
cryptography>=42.0.0

# Data analyst sandbox
docker>=7.0.0
neo4j>=5.23.0

# Data analyst tools
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
openpyxl>=3.1.0
xlrd>=2.0.0
